Implement prepared statements which return data for MySQL
After this change, the core crate's unit test suite is 100% green for
MySQL (`cd diesel && cargo test --features mysql`)

So here's how loading data works. Output values use the same data
structure as input values. For types with a fixed size, we can allocate
the small buffer needed, and everything will just work. For things that
we don't know the size of until we fetch the row, we set the buffer to
size 0.

When mysql hits a field where the buffer isn't large enough, the return
value of `mysql_stmt_fetch` will be `MYSQL_DATA_TRUNCATED`. For each
field that was truncated, the value of `*error` will be set to non-zero
(it's a `my_bool`, so presumably that value is 1. This appears to be the
only error. I do not know why they didn't call it `is_truncated`). It
will also set `*length` to the number of bytes required to hold the
value. At this point we can reallocate the buffer, and try to load that
column directly.

Aside from the general horribleness of this API that I've ranted about
elsewhere, two things in particular greatly bother me here:

- `MYSQL_DATA_TRUNCATED` and all other values mentioned as return values
  for `mysql_stmt_fetch` are `unsigned int`. `mysql_stmt_fetch` returns
  `int`.
- The MySQL client C API is clearly buffering this value on its own. Why
  the fuck can it not just give me that pointer, a byte count, and tell
  me to `memcpy` it before calling `mysql_stmt_fetch` again if I need
  it?!

Probably the worst part of this code is the implementation of
`StatementIterator::next`. Since we're actually just mutating the same
buffers over and over, I need to yield the same value each time. Of
course I can't just return `Self`, so I need to return a reference to
the appropriate value. There's not really any way to do this without
unsafe code. In particular what I need is a way to guarantee that the
value is unused by the next call to `next`, but I have no way to express
that. So I'm transmuting the reference to `'static`, and never exposing
this API. I know it is safe in the place this is currently used. In
theory we could at least ensure no use after free by implementing
iterator on a reference instead of the struct, so we have a smaller
lifetime. Even if we did that though, we're still mutating immutable
data, which is UB. So given that any code which holds onto that
reference past the next call to `next` is invoking UB no matter what, I
went with the smallest possible solution. I am open to recommendations
here.